cleaned up a bit

+++ src/mandatory/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: lverdoes <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/30 14:17:12 by lverdoes      #+#    #+#                 */
-/*   Updated: 2021/02/26 15:05:59 by lverdoes      ########   odam.nl         */
+/*   Updated: 2021/06/30 22:42:35 by lverdoes      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_atoi(const char *str)
 {
-	unsigned long long	result;
+	size_t	result;
 	int					neg;
 	size_t				i;
